handle projects settings
